crossref-django/requirements.txt
Django>=4.2,<5
djangorestframework
httpx
python-dotenv
gunicorn
